Remove redundant parent filter on full union pushdown
When the optimizer successfully pushes a timestamp filter into every
branch of a UNION/UNION ALL/INTERSECT/EXCEPT, it now removes the
parent-level copy of that filter. Previously it always retained the
parent filter as a safety net, causing an unnecessary re-check on the
combined result set.

Three production changes in SqlOptimiser:
- tryPushFilterIntoSetOperationBranches returns a boolean (true when
  all branches received the filter), and the call site keeps the
  parent filter only when pushdown was incomplete.
- The call site evaluates the union check before the blanket barrier
  check (LATEST ON / LIMIT / non-pushable SAMPLE BY), so UNION models
  with a barrier on the first branch still enter per-branch pushdown.
- findTimestamp iterates branches to resolve the timestamp alias when
  the first branch obscures it (e.g. SAMPLE BY wraps it in
  timestamp_floor), mapping back by column position.

All filter-pushdown-into-union tests move into the new
FilterPushdownIntoUnionTest file, including four new partial-pushdown
tests that verify the optimizer retains the parent filter when
SAMPLE BY blocks pushdown on specific branches.
