+++ core/src/test/java/io/questdb/test/griffin/SqlParserTest.java
@@ -9796,13 +9796,11 @@ public class SqlParserTest extends AbstractSqlParserTest {
     public void testPushDownTimestampFilterThroughExcept() throws SqlException {
         assertQuery(
                 "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") except select-choose [ts2 ts] ts2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
+                        /**/ "select-choose [ts1 ts] ts1 ts from (" +
+                        /**/   "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
+                        /**/ ") except select-choose [ts2 ts] ts2 ts from (" +
+                        /**/   "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
+                        /**/ ")" +
                         ")",
                 "select ts from (select ts1 ts from t1 except select ts2 ts from t2) where ts in '2025-12-01T01;2h'",
                 modelOf("t1").timestamp("ts1"),
@@ -9814,252 +9812,15 @@ public class SqlParserTest extends AbstractSqlParserTest {
     public void testPushDownTimestampFilterThroughIntersect() throws SqlException {
         assertQuery(
                 "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") intersect select-choose [ts2 ts] ts2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
+                        /**/ "select-choose [ts1 ts] ts1 ts from (" +
+                        /**/   "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
+                        /**/ ") intersect select-choose [ts2 ts] ts2 ts from (" +
+                        /**/   "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
+                        /**/ ")" +
                         ")",
                 "select ts from (select ts1 ts from t1 intersect select ts2 ts from t2) where ts in '2025-12-01T01;2h'",
                 modelOf("t1").timestamp("ts1"),
                 modelOf("t2").timestamp("ts2")
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnion() throws SqlException {
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") union select-choose [ts2 ts] ts2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                "select ts from (select ts1 ts from t1 union select ts2 ts from t2) where ts in '2025-12-01T01;2h'",
-                modelOf("t1").timestamp("ts1"),
-                modelOf("t2").timestamp("ts2")
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnionAll() throws SqlException {
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-choose [ts2 ts] ts2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                "select ts from (select ts1 ts from t1 union all select ts2 ts from t2) where ts in '2025-12-01T01;2h'",
-                modelOf("t1").timestamp("ts1"),
-                modelOf("t2").timestamp("ts2")
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnionAllCte() throws SqlException {
-        // Two separate CTEs combined with UNION ALL in the main query.
-        // The timestamp filter must be pushed into both CTE branches.
-        createModelsAndRun(
-                () -> {
-                    try (SqlCompiler compiler = engine.getSqlCompiler()) {
-                        // First verify that a single CTE wrapping the whole UNION works
-                        // (structurally identical to the nested subquery form)
-                        sink.clear();
-                        ExecutionModel singleCteModel = compiler.generateExecutionModel(
-                                "WITH u AS (SELECT ts1 ts FROM t1 UNION ALL SELECT ts2 ts FROM t2) " +
-                                        "SELECT ts FROM u WHERE ts in '2025-12-01T01;2h'",
-                                sqlExecutionContext);
-                        ((Sinkable) singleCteModel).toSink(sink);
-                        String singleCtePlan = sink.toString();
-
-                        Assert.assertTrue(
-                                "single CTE: timestamp filter not pushed into first branch: " + singleCtePlan,
-                                singleCtePlan.contains("where ts1 in '2025-12-01T01;2h'"));
-                        Assert.assertTrue(
-                                "single CTE: timestamp filter not pushed into second branch: " + singleCtePlan,
-                                singleCtePlan.contains("where ts2 in '2025-12-01T01;2h'"));
-
-                        // Now verify that two separate CTEs combined with UNION ALL also push down
-                        sink.clear();
-                        ExecutionModel twoCteModel = compiler.generateExecutionModel(
-                                "WITH l AS (SELECT ts1 ts FROM t1), " +
-                                        "r AS (SELECT ts2 ts FROM t2) " +
-                                        "SELECT ts FROM (l UNION ALL r) " +
-                                        "WHERE ts in '2025-12-01T01;2h'",
-                                sqlExecutionContext);
-                        ((Sinkable) twoCteModel).toSink(sink);
-                        String twoCtePlan = sink.toString();
-
-                        Assert.assertTrue(
-                                "two CTEs: timestamp filter not pushed into first CTE branch: " + twoCtePlan,
-                                twoCtePlan.contains("where ts1 in '2025-12-01T01;2h'"));
-                        Assert.assertTrue(
-                                "two CTEs: timestamp filter not pushed into second CTE branch: " + twoCtePlan,
-                                twoCtePlan.contains("where ts2 in '2025-12-01T01;2h'"));
-                    }
-                },
-                modelOf("t1").timestamp("ts1"),
-                modelOf("t2").timestamp("ts2")
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnionAllMismatchedAliases() throws SqlException {
-        // The "ts" alias maps to the TIMESTAMP column in branch 1 (position 1),
-        // but to the SYMBOL column in branch 2 (position 2). The filter should be
-        // remapped by position so that branch 2 filters on ts2 (the TIMESTAMP at
-        // position 1), not on sym2 which carries the "ts" alias.
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] name1, ts1 ts, sym1 from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-choose [ts2] name2, ts2, sym2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                """
-                        SELECT ts FROM (
-                            SELECT name1, ts1 ts, sym1 FROM t1
-                            UNION ALL
-                            SELECT name2, ts2, sym2 ts FROM t2
-                        ) WHERE ts IN '2025-12-01T01;2h'
-                        """,
-                modelOf("t1").col("name1", ColumnType.VARCHAR).timestamp("ts1").col("sym1", ColumnType.SYMBOL),
-                modelOf("t2").col("name2", ColumnType.VARCHAR).timestamp("ts2").col("sym2", ColumnType.SYMBOL)
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnionAllNonPushableBranch() throws SqlException {
-        // One branch aliases the column to a non-literal expression (ts2+1).
-        // The filter is pushed to both branches (since the parent's alias map sees it as a literal),
-        // but on the non-literal branch it stays at the select-virtual level and cannot be pushed further.
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-virtual [ts2 + 1 ts] ts2 + 1 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2)" +
-                        /**/   ") where ts in '2025-12-01T01;2h'" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                "select ts from (select ts1 ts from t1 union all select ts2+1 ts from t2) where ts in '2025-12-01T01;2h'",
-                modelOf("t1").timestamp("ts1"),
-                modelOf("t2").timestamp("ts2")
-        );
-    }
-
-    @Test
-    public void testPushDownTimestampFilterThroughUnionAllThreeBranches() throws SqlException {
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts1 ts] ts1 ts from (" +
-                        /**/     "select [ts1] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-choose [ts2 ts] ts2 ts from (" +
-                        /**/     "select [ts2] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-choose [ts3 ts] ts3 ts from (" +
-                        /**/     "select [ts3] from t3 timestamp (ts3) where ts3 in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                "select ts from (select ts1 ts from t1 union all select ts2 ts from t2 union all select ts3 ts from t3) where ts in '2025-12-01T01;2h'",
-                modelOf("t1").timestamp("ts1"),
-                modelOf("t2").timestamp("ts2"),
-                modelOf("t3").timestamp("ts3")
-        );
-    }
-
-    @Test
-    public void testPushFilterThroughUnionAllExprInAllBranches() throws SqlException {
-        // Non-timestamp filter is not pushed into union branches.
-        assertQuery(
-                "select-choose c from (" +
-                        /**/ "select [c] from (" +
-                        /**/   "select-virtual [x + y c] x + y c from (" +
-                        /**/     "select [y, x] from t1" +
-                        /**/   ") union all select-virtual [x + y c] x + y c from (" +
-                        /**/     "select [y, x] from t2" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where c > 5" +
-                        ")",
-                "select c from (select x + y c from t1 union all select x + y c from t2) where c > 5",
-                modelOf("t1").col("x", ColumnType.INT).col("y", ColumnType.INT),
-                modelOf("t2").col("x", ColumnType.INT).col("y", ColumnType.INT)
-        );
-    }
-
-    @Test
-    public void testPushFilterThroughUnionAllMixedPushability() throws SqlException {
-        // Two filters: ts is the designated timestamp (pushed into branches for partition pruning),
-        // c is a non-timestamp expression (stays at parent level only).
-        assertQuery(
-                "select-choose ts, c from (" +
-                        /**/ "select [ts, c] from (" +
-                        /**/   "select-virtual [ts, x + y c] ts, x + y c from (" +
-                        /**/     "select-choose [ts1 ts, y, x] ts1 ts, y, x from (" +
-                        /**/       "select [ts1, y, x] from t1 timestamp (ts1) where ts1 in '2025-12-01T01;2h'" +
-                        /**/     ")" +
-                        /**/   ")" +
-                        /**/   " union all select-virtual [ts, x + y c] ts, x + y c from (" +
-                        /**/     "select-choose [ts2 ts, y, x] ts2 ts, y, x from (" +
-                        /**/       "select [ts2, y, x] from t2 timestamp (ts2) where ts2 in '2025-12-01T01;2h'" +
-                        /**/     ")" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h' and c > 5" +
-                        ")",
-                "select ts, c from (select ts1 ts, x + y c from t1 union all select ts2 ts, x + y c from t2) where ts in '2025-12-01T01;2h' and c > 5",
-                modelOf("t1").timestamp("ts1").col("x", ColumnType.INT).col("y", ColumnType.INT),
-                modelOf("t2").timestamp("ts2").col("x", ColumnType.INT).col("y", ColumnType.INT)
-        );
-    }
-
-    @Test
-    public void testPushFilterThroughUnionAllSameTable() throws SqlException {
-        // Both branches reference the same table; filter should be pushed to both independently.
-        assertQuery(
-                "select-choose ts from (" +
-                        /**/ "select [ts] from (" +
-                        /**/   "select-choose [ts] ts from (" +
-                        /**/     "select [ts] from t1 timestamp (ts) where ts in '2025-12-01T01;2h'" +
-                        /**/   ") union all select-choose [ts] ts from (" +
-                        /**/     "select [ts] from t1 timestamp (ts) where ts in '2025-12-01T01;2h'" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts in '2025-12-01T01;2h'" +
-                        ")",
-                "select ts from (select ts from t1 union all select ts from t1) where ts in '2025-12-01T01;2h'",
-                modelOf("t1").timestamp("ts")
-        );
-    }
-
-    @Test
-    public void testPushFilterWithMultipleColumnsThroughUnionAll() throws SqlException {
-        // Filter references two columns (ts and v); not pushed because it references a non-timestamp column.
-        assertQuery(
-                "select-choose ts, v from (" +
-                        /**/ "select [ts, v] from (" +
-                        /**/   "select-choose [val1 v, ts1 ts] ts1 ts, val1 v from (" +
-                        /**/     "select [val1, ts1] from t1" +
-                        /**/   ") union all select-choose [val2 v, ts2 ts] ts2 ts, val2 v from (" +
-                        /**/     "select [val2, ts2] from t2" +
-                        /**/   ")" +
-                        /**/ ") _xQdbA1 where ts > v" +
-                        ")",
-                "select ts, v from (select ts1 ts, val1 v from t1 union all select ts2 ts, val2 v from t2) where ts > v",
-                modelOf("t1").col("ts1", ColumnType.TIMESTAMP).col("val1", ColumnType.INT),
-                modelOf("t2").col("ts2", ColumnType.TIMESTAMP).col("val2", ColumnType.INT)
         );
     }
